+++ blobs.js
@@ -2,14 +2,17 @@ let blobManager;
 let manualInput;
 let joystick;
 let player;
+let followCam;
 let gameboard;
-let gamesize = 2000;
+let gamesize = 1000;
 
 function setup() {
 	let canvas = createCanvas(
-		window.innerWidth /1.5,
-		window.innerHeight /1.5
+		window.innerWidth * 5/7,
+		window.innerHeight * 5/7
 		);
+
+	// frameRate(1);
 
 	canvas.parent('sketch-holder');
 
@@ -18,6 +21,8 @@ function setup() {
 	manualInput = new ManualInput(joystick);
 	blobManager = new BlobManager(gameboard);
 	player = new Player(blobManager, manualInput);
+	followCam = new FollowCam(blobManager);
+	followCam.follow(player);
 }
 
 function draw() {
@@ -26,4 +31,5 @@ function draw() {
 	blobManager.update();
 	manualInput.update();
 	joystick.run();
+	followCam.update();
 }