+++ blobmanager.js
@@ -1,11 +1,13 @@
 class BlobManager {
-	constructor() {
+	constructor(gameboard) {
+		this.gameboard = gameboard;
+
 		this.blobs = [];
-		this.initialBlobAmount = 6;
+		this.initialBlobAmount = 20;
 		this.standardBlobSize = 20;
 
 		colorMode(HSB, 255, 255, 255);
-		this.playerBlobColor = color(0, 255, 170);
+		this.playerBlobColor = color(1, 255, 170);
 
 		this.initBlobs();
 
@@ -13,10 +15,13 @@ class BlobManager {
 	}
 
 	update() {
+		let thisHandle = this;
 		this.blobs.forEach(function(blob) {
-			blob.run();
+			blob.update();
+			thisHandle.repositionOutsideGameboard(blob);
 		});
 		this.checkForCollisions();
+
 		this.deleteDeadBlobs();
 	}
 
@@ -27,7 +32,7 @@ class BlobManager {
 	}
 
 	addBlob() {
-		let pos = createVector(random(width), random(height));
+		let pos = createVector(random(this.gameboard.width), random(this.gameboard.height));
 		let vel = createVector(randomGaussian(0, 2), randomGaussian(0, 2));
 		let size = randomGaussian(this.standardBlobSize, 0.1);
 		let col = color(random(255), 120, 230);
@@ -46,6 +51,25 @@ class BlobManager {
 		this.playerBlob = new Blob(size, this.playerBlobColor, pos, vel, isManual);
 		this.blobs.push(this.playerBlob);
 		return this.playerBlob;
+	}
+
+	repositionOutsideGameboard(blob) {
+		if (blob.pos.x > this.gameboard.width) {
+			blob.pos.x = 0;
+		}
+		if(blob.pos.x < 0) {
+			blob.pos.x = this.gameboard.width;
+		}
+		if (blob.pos.y > this.gameboard.height) {
+			blob.pos.y = 0;
+		}
+		if (blob.pos.y < 0) {
+			blob.pos.y = this.gameboard.height;
+		}
+	}
+
+	displayThoseWithinView(topLeft, bottomRight) {
+		
 	}
 
 	checkForCollisions() {
@@ -83,5 +107,9 @@ class BlobManager {
 			let index = this.blobs.indexOf(deadBlob);
 			this.blobs.splice(index, 1);
 		}
+	}
+
+	get allBlobs() {
+		return this.blobs;
 	}
 }