+++ blobs.js
@@ -2,19 +2,27 @@ let blobManager;
 let manualInput;
 let joystick;
 let player;
+let followCam;
+let gameboard;
+let gamesize = 1000;
 
 function setup() {
 	let canvas = createCanvas(
-		window.innerWidth /1.5,
-		window.innerHeight /1.5
+		window.innerWidth * 5/7,
+		window.innerHeight * 5/7
 		);
+
+	// frameRate(1);
 
 	canvas.parent('sketch-holder');
 
+	gameboard = new Gameboard(gamesize, gamesize);
 	joystick = new Joystick();
 	manualInput = new ManualInput(joystick);
-	blobManager = new BlobManager();
+	blobManager = new BlobManager(gameboard);
 	player = new Player(blobManager, manualInput);
+	followCam = new FollowCam(blobManager);
+	followCam.follow(player);
 }
 
 function draw() {
@@ -23,4 +31,5 @@ function draw() {
 	blobManager.update();
 	manualInput.update();
 	joystick.run();
+	followCam.update();
 }