+++ gameboard.js
@@ -1,0 +1,6 @@
+class Gameboard {
+	constructor(boardWidth, boardHeight) {
+		this.width = boardWidth;
+		this.height = boardHeight;
+	}
+}