+++ particle.js
@@ -14,14 +14,16 @@ class Particle {
 		this.vel.limit(this.maxSpeed);
 		this.pos.add(this.vel);
 		this.acc.mult(0);
-		
-		this.edges();
 	}
 
 	display() {
+		displayAt(this.pos);
+	}
+
+	displayAt(pos) {
 		noStroke();
 		fill(this.color);
-		ellipse(this.pos.x, this.pos.y, this.size*2);
+		ellipse(pos.x, pos.y, this.size*2);
 	}
 
 	run() {
@@ -37,20 +39,5 @@ class Particle {
 
 	get mass() {
 		return this.size;
-	}
-
-	edges() {
-		if (this.pos.x > width) {
-			this.pos.x = 0;
-		}
-		if(this.pos.x < 0) {
-			this.pos.x = width;
-		}
-		if (this.pos.y > height) {
-			this.pos.y = 0;
-		}
-		if (this.pos.y < 0) {
-			this.pos.y = height;
-		}
 	}
 }